backend: add unittests for the publisher
All preparations have been made to make sure we can test the publisher.
Now add tests to verify the behavior of the publisher. We send messages,
connections, and then inspect what messages are forwarded.

The stepChan is used to make sure everything is executed in the expected
order to make the system behave predictable under test.