backend: fix empty error messages
This is broken because json.Unmarshal does not handle empty strings. The
issue was hidden because logging the error did not work because the
final `.Msg()` method call was missing, meaning nothing was actually
logged.

Because of this, empty error messages, which indicate any previous
error has been solved, were sent as empty `GenericMessage`.