frontend: sort servers from new to old
Now, the sort order is from old to new. Because most changes happen on
newer branches, it makes sense to invert the order.

But we want order per version still in alphabetical descending order, so
we first reverse sort on the version, and if they are equal, sort the
architecture.